+++ Common/Button.h
@@ -20,6 +20,6 @@ public:
 	virtual void implement_function() {};
 	void mousePressed(int x, int y, bool isLeft);
 	void expand_string(string &s);
-	void draw(Graphics& g, int x, int y, size_t z);
+	void draw(Graphics& g, short x, short y, size_t z);
 	virtual void setValue(string s) ;
 };
